+++ pkg/testutil/file.go
@@ -1,7 +1,9 @@
 package testutil
 
 import (
+	"io"
 	"os"
+	"path/filepath"
 	"testing"
 )
 
@@ -17,4 +19,37 @@ func MakeTempFile(t *testing.T, contents string) string {
 		t.Fatal(err)
 	}
 	return tmpfile.Name()
+}
+
+// CopyFile copies a referenced file to a new location.
+func CopyFile(t *testing.T, fromFilename, toFilename string) {
+	t.Helper()
+
+	src, err := os.Open(fromFilename)
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer src.Close()
+
+	toDir := filepath.Dir(toFilename)
+	if err := os.MkdirAll(toDir, 0777); err != nil {
+		t.Fatal(err)
+	}
+
+	dst, err := os.Create(toFilename)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if _, err := io.Copy(dst, src); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := dst.Sync(); err != nil {
+		t.Fatal(err)
+	}
+
+	if err := dst.Close(); err != nil {
+		t.Fatal(err)
+	}
 }
